# docs/MOBILE_V1_OPENAPI.yaml
openapi: 3.0.3
info:
  title: MESC Mobile API
  version: 1.0.0
  description: Initial mobile-native contract for MESC Native MVP milestone 1.
servers:
  - url: /api/mobile/v1
security:
  - bearerAuth: []
paths:
...
  /notifications:
    get:
      summary: List recent notifications for the authenticated user.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Notification list.
          content:
            application/json:
              schema:
                type: object
                required: [success, notifications, unreadCount]
                properties:
                  success:
                    type: boolean
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/MobileNotification"
                  unreadCount:
                    type: integer
...
  /notifications/{id}/read:
    patch:
      summary: Mark one notification as read.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Notification marked as read.
          content:
            application/json:
              schema:
                type: object
                required: [success, notification]
                properties:
                  success:
                    type: boolean
                  notification:
                    type: object
                    required: [id, read, readAt]
                    properties:
                      id:
                        type: string
                      read:
                        type: boolean
                      readAt:
                        type: string
                        nullable: true
        "404":
          description: Notification not found for the current user.
...
  /schedules/{id}:
    get:
      summary: Read a single schedule visible to the authenticated user.
      parameters:
        - $ref: "#/components/parameters/CommunityHeader"
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Schedule details.
          content:
            application/json:
              schema:
                type: object
                required: [success, schedule]
                properties:
                  success:
                    type: boolean
                  schedule:
                    $ref: "#/components/schemas/ScheduleDetail"
        "403":
          description: Schedule is outside the user's permission scope.
        "404":
          description: Schedule not found.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CommunityHeader:
      in: header
      name: X-Community-Id
      schema:
        type: string
      required: false
      description: Optional active community. Community-scoped users can only request their home community.
...
  schemas:
...
    MobileNotification:
      type: object
      required: [id, type, eventKey, title, message, priority, read, readAt, deepLink, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
        eventKey:
          type: string
          nullable: true
          enum:
            - questionnaire_published
            - coordinator_announcement
            - questionnaire_closed
            - schedule_published
            - substitution_requested
            - substitute_accepted
            - formation_available
            - schedule_reminder
        title:
          type: string
        message:
          type: string
        priority:
          type: string
          nullable: true
        read:
          type: boolean
        readAt:
          type: string
          nullable: true
        deepLink:
          type: string
        createdAt:
          type: string
          nullable: true
...
    ScheduleSummary:
      type: object
      required: [id, date, time, type, position, status, deepLink]
      properties:
        id:
          type: string
        date:
          type: string
          nullable: true
        time:
          type: string
        type:
          type: string
        location:
          type: string
          nullable: true
        position:
          type: integer
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
        deepLink:
          type: string
    ScheduleDetail:
      allOf:
        - $ref: "#/components/schemas/ScheduleSummary"
        - type: object
          properties:
            confirmationStatus:
              type: string
              nullable: true
            substitution:
              type: object
              nullable: true
            canConfirm:
              type: boolean
            canRequestSubstitution:
              type: boolean
...